+++ keyboards/dztech/dz65rgb/keymaps/kgreulich/config.h
@@ -10,9 +10,6 @@
 //#define QMK_KEYS_PER_SCAN 4
 
 // some speed shit
-#define NO_ACTION_MACRO
-#define NO_ACTION_FUNCTION
-
 #ifndef NO_DEBUG
 #define NO_DEBUG
 #endif // !NO_DEBUG
@@ -24,13 +21,9 @@
 #define TAPPING_FORCE_HOLD
 #define IGNORE_MOD_TAP_INTERRUPT
 
-// #include "config_led.h"
-// #include "dz60rgb.h"
-#undef DISABLE_RGB_MATRIX_SPLASH
-#undef DISABLE_RGB_MATRIX_MULTISPLASH
-#undef DISABLE_RGB_MATRIX_SOLID_MULTISPLASH
 #undef RGB_MATRIX_STARTUP_MODE
-#define RGB_MATRIX_FRAMEBUFFER_EFFECTS
+#undef RGB_MATRIX_FRAMEBUFFER_EFFECTS
+
 #define DISABLE_RGB_MATRIX_SOLID_COLOR
 #define DISABLE_RGB_MATRIX_ALPHAS_MODS
 #define DISABLE_RGB_MATRIX_GRADIENT_UP_DOWN
@@ -56,8 +49,8 @@
 #define DISABLE_RGB_MATRIX_JELLYBEAN_RAINDROPS
 #define DISABLE_RGB_MATRIX_TYPING_HEATMAP
 #define DISABLE_RGB_MATRIX_DIGITAL_RAIN
-//#define DISABLE_RGB_MATRIX_SOLID_REACTIVE_SIMPLE
-//#define DISABLE_RGB_MATRIX_SOLID_REACTIVE
+#define DISABLE_RGB_MATRIX_SOLID_REACTIVE_SIMPLE
+#define DISABLE_RGB_MATRIX_SOLID_REACTIVE
 #define DISABLE_RGB_MATRIX_SOLID_REACTIVE_WIDE
 #define DISABLE_RGB_MATRIX_SOLID_REACTIVE_MULTIWIDE
 #define DISABLE_RGB_MATRIX_SOLID_REACTIVE_CROSS
