+++ keyboards/dztech/dz65rgb/keymaps/kgreulich/rules.mk
@@ -1,0 +1,6 @@
+NKRO_ENABLE = yes                 # USB Nkey Rollover
+
+LINK_TIME_OPTIMIZATION_ENABLE = yes
+
+RGBLIGHT_ENABLE = no
+
